scripts: avoid error when running on dynamically started VM
If testing is run on a pair of dyanamically started VMs, there
is no TE_IUT and no TE_TST1  when TE is starting. So, we can't
check if a module is loaded on not. Keep list of modules empty
and, therefore, assume that it is not loaded.

Ideally we should get rid of it completely since it is unfriendly
to testing on dynamically started VMs. Above assumption could be
and is wrong in virtio-pci case, but it does not matter a lot.

Signed-off-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>