ool/config: check conditions to run af_xdp without filters
af_xdp_no_filetrs should be tested using the interface that is
configured with a single combined channel, because multiple
channels without set flow filter may lead the af_xdp socket
get only part of the trafic. To avoid such misconfigured test cases,
we check the number of combined channels on the aim interface
and exit with failure if the number is not equal one.

OL-Redmine-Id: 12237
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Acked-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>