scripts: remove sudo when delete libte-iut.so
Not all users may be permitted to launch commands via sudo
without additional password verification. Lack of the permission
may be a cause of appearance of unwanted messages about necessity
to read the password by terminal. In l5_run script, we have removed
sudo from the command deleting libte-iut.so library on iut host to
prevent appearing of such redundant messages for users without
the privilege.

OL-Redmine-Id: 12650
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>