scripts/l5_run: fix getting default EF_* variable values
Fix find_ef_default() function for the cases when we try to get the
default value of non-existent variable.
Tell 'sed' to print only matched lines and omit all others. Otherwise,
if there is no occurrences, it will print out all the input and put
invalid data as a result.

OL-Redmine-Id: 12616
Fixes: 28e1f46bad27 ("ool/config: set EF variable to default value")
Signed-off-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>