cs/socktest: enable stack allocation failure handler back
Failed to allocate stack is usefull instrument to spot errors
on systems with limited amount of stacks. Some tests are having
SIGUSR2 from failed to allocate stack pattern in theirs verdicts.
Most of this error is already fixed and if error with "Failed to
allocate stack" occur, it should be fixed with newly added function
to kill zombie stacks, but not ignored. So commit 5414d76d was reverted.

Fixes: 5414d76d ("cs/socktest.yml: remove the storm of unnecessary errors")
OL-Redmine-Id: 12903
Signed-off-by: Demid Trempolskii <demid.trempolskii@arknetworks.am>
Reviewed-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>