scripts/lib: use the TE mechanism to run commands on TA
There are still some functions in the file that cannot yet be modified
to use TE's mechanism, because they operate with the host name, not TA.
In the future, this will also need to be fixed when it is required in
any TS (and it will probably require synchronous changes in all test
suites involved). Now it will continue to work the same way as before.

OL-Redmine-Id: 13706
Signed-off-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>